function_app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry
azure-functions
#azure-functions==1.21.0
requests
azure-identity
azure-keyvault-secrets
#cryptography==41.0.7
sqlalchemy
psycopg2-binary
